# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hopfield Assignment Problem Solver API
  description: |
    A high-performance API service for solving assignment problems using Hopfield neural networks.
    
    The assignment problem consists of assigning n jobs to n workers such that the total cost is minimized.
    
    ## Authentication
    API requests require an API key to be provided in one of the following ways:
    - Header: `X-API-Key: your-api-key`
    - Header: `Authorization: Bearer your-api-key`
    
    ## Rate Limiting
    - API Gateway: 100 requests per minute
    - Requests exceeding the limit will receive a 429 status code
    
    ## Constraints
    - Matrix size: 2x2 to 50x50
    - Cost values: 0 to 1,000,000,000
    - Batch size: Maximum 100 problems per batch
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/your-repo
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development server (API Gateway)
  - url: http://localhost
    description: Production server (via Nginx)
  - url: http://localhost:5000
    description: Hopfield Service (direct access)

tags:
  - name: health
    description: Service health and status endpoints
  - name: solve
    description: Assignment problem solving endpoints
  - name: metrics
    description: Metrics and monitoring endpoints

paths:
  /health:
    get:
      tags:
        - health
      summary: Health check
      description: Check if the service is healthy
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /health/ready:
    get:
      tags:
        - health
      summary: Readiness check
      description: Check if the service is ready to accept requests
      operationId: readyCheck
      responses:
        '200':
          description: Service is ready
        '503':
          description: Service is not ready

  /health/live:
    get:
      tags:
        - health
      summary: Liveness check
      description: Check if the service is alive
      operationId: liveCheck
      responses:
        '200':
          description: Service is alive

  /solve:
    post:
      tags:
        - solve
      summary: Solve an assignment problem
      description: Solve a single assignment problem using Hopfield neural network
      operationId: solveAssignment
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolveRequest'
            examples:
              simple:
                summary: Simple 2x2 problem
                value:
                  cost_matrix:
                    - [1, 2]
                    - [3, 4]
              standard:
                summary: Standard 4x4 problem
                value:
                  cost_matrix:
                    - [9, 2, 7, 8]
                    - [6, 4, 3, 7]
                    - [5, 8, 1, 8]
                    - [7, 6, 9, 4]
      responses:
        '200':
          description: Problem solved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolveResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /solve/batch:
    post:
      tags:
        - solve
      summary: Solve multiple assignment problems
      description: Solve multiple assignment problems in batch
      operationId: solveBatch
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRequest'
            examples:
              sample:
                summary: Sample batch request
                value:
                  problems:
                    - id: "problem_1"
                      cost_matrix:
                        - [1, 2]
                        - [3, 4]
                    - id: "problem_2"
                      cost_matrix:
                        - [5, 6, 7]
                        - [8, 9, 10]
                        - [11, 12, 13]
      responses:
        '200':
          description: Batch processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /metrics:
    get:
      tags:
        - metrics
      summary: Get service metrics
      description: Get current service metrics and statistics
      operationId: getMetrics
      responses:
        '200':
          description: Metrics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          example: healthy
        service:
          type: string
          example: hopfield-assignment-api
        version:
          type: string
          example: "1.0.0"
        timestamp:
          type: string
          format: date-time

    SolveRequest:
      type: object
      required:
        - cost_matrix
      properties:
        cost_matrix:
          type: array
          description: Square cost matrix (n x n)
          minItems: 2
          maxItems: 50
          items:
            type: array
            items:
              type: number
              format: float
              minimum: 0
              maximum: 1000000000
          example:
            - [1, 2]
            - [3, 4]

    SolveResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        result:
          type: object
          properties:
            assignments:
              type: array
              description: Assignment results (assignments[i] = job for worker i)
              items:
                type: integer
              example: [0, 1]
            total_cost:
              type: number
              format: float
              description: Total cost of the optimal assignment
              example: 5.0
            iterations:
              type: integer
              description: Number of iterations to convergence
              example: 45
            cost_matrix:
              type: array
              description: Original cost matrix
              items:
                type: array
                items:
                  type: number

    BatchRequest:
      type: object
      required:
        - problems
      properties:
        problems:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required:
              - id
              - cost_matrix
            properties:
              id:
                type: string
                description: Unique identifier for this problem
                example: "problem_1"
              cost_matrix:
                type: array
                items:
                  type: array
                  items:
                    type: number

    BatchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              success:
                type: boolean
              result:
                $ref: '#/components/schemas/SolveResponse'
              error:
                type: string

    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Field 'cost_matrix' is required"
        request_id:
          type: string
          format: uuid
          description: Unique request identifier for debugging

    MetricsResponse:
      type: object
      properties:
        requests:
          type: object
          properties:
            total:
              type: integer
            errors:
              type: integer
            success_rate:
              type: number
              format: float
        performance:
          type: object
          properties:
            avg_duration_ms:
              type: number
            min_duration_ms:
              type: number
            max_duration_ms:
              type: number
        algorithm:
          type: object
          properties:
            avg_iterations:
              type: number
            avg_matrix_size:
              type: number